+++ example/spec/parallel/ingestGranule/IngestGranuleFtpSuccessSpec.js
@@ -10,10 +10,10 @@ const {
   api: apiTestUtils,
   buildAndExecuteWorkflow,
   cleanupCollections,
-  granulesApi: granulesApiTestUtils,
 } = require('@cumulus/integration-tests');
-const { deleteProvider } = require('@cumulus/api-client/providers');
 const { deleteExecution } = require('@cumulus/api-client/executions');
+const { getGranule, deleteGranule } = require('@cumulus/api-client/granules');
+const { deleteProvider } = require('@cumulus/api-client/providers');
 const mime = require('mime-types');
 const { loadConfig, createTimestampedTestId, createTestSuffix } = require('../../helpers/testUtils');
 const { waitForModelStatus } = require('../../helpers/apiUtils');
@@ -99,7 +99,7 @@ describe('The FTP Ingest Granules workflow', () => {
         'completed'
       );
 
-      granuleResponse = await granulesApiTestUtils.getGranule({
+      granuleResponse = await getGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId,
       });
@@ -108,7 +108,7 @@ describe('The FTP Ingest Granules workflow', () => {
 
     afterAll(async () => {
       // clean up granule
-      await granulesApiTestUtils.deleteGranule({
+      await deleteGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId,
       });