+++ example/spec/parallel/syncGranule/SyncGranuleSuccessSpec.js
@@ -1,20 +1,19 @@
 const fs = require('fs');
 const difference = require('lodash/difference');
 const {
-  buildAndExecuteWorkflow,
+  addCollections,
   addProviders,
   api: apiTestUtils,
-  cleanupProviders,
-  addCollections,
+  buildAndExecuteWorkflow,
   cleanupCollections,
-  granulesApi: granulesApiTestUtils,
-  waitForTestExecutionStart,
+  cleanupProviders,
   waitForCompletedExecution,
+  waitForTestExecutionStart,
 } = require('@cumulus/integration-tests');
 const { updateCollection } = require('@cumulus/integration-tests/api/api');
 const { Execution, Granule } = require('@cumulus/api/models');
 const { deleteExecution } = require('@cumulus/api-client/executions');
-const { deleteGranule } = require('@cumulus/api-client/granules');
+const { deleteGranule, getGranule, reingestGranule } = require('@cumulus/api-client/granules');
 const { s3 } = require('@cumulus/aws-client/services');
 const {
   s3GetObjectTagging,
@@ -51,16 +50,16 @@ describe('The Sync Granules workflow', () => {
   let executionModel;
   let expectedPayload;
   let expectedS3TagSet;
+  let failingExecutionArn;
   let granuleModel;
   let inputPayload;
   let lambdaStep;
   let provider;
+  let reingestGranuleExecutionArn;
+  let syncGranuleExecutionArn;
   let testDataFolder;
   let testSuffix;
   let workflowExecution;
-  let syncGranuleExecutionArn;
-  let reingestGranuleExecutionArn;
-  let failingExecutionArn;
 
   beforeAll(async () => {
     config = await loadConfig();
@@ -153,15 +152,15 @@ describe('The Sync Granules workflow', () => {
 
   afterAll(async () => {
     // clean up stack state added by test
-    await deleteGranule({
-      prefix: config.stackName,
-      granuleId: inputPayload.granules[0].granuleId,
-    });
-
-    await apiTestUtils.deletePdr({
+    await Promise.all(inputPayload.granules.map(
+      (granule) => deleteGranule({
+        prefix: config.stackName,
+        granuleId: granule.granuleId,
+      })
+    )).then(() => apiTestUtils.deletePdr({
       prefix: config.stackName,
       pdr: inputPayload.pdr.name,
-    });
+    }));
 
     await Promise.all([
       deleteExecution({ prefix: config.stackName, executionArn: syncGranuleExecutionArn }),
@@ -258,7 +257,7 @@ describe('The Sync Granules workflow', () => {
     });
   });
 
-  describe('the reporting lambda has received the cloudwatch stepfunction event and', () => {
+  describe('the reporting lambda has received the CloudWatch step function event and', () => {
     it('the execution record is added to DynamoDB', async () => {
       const record = await waitForModelStatus(
         executionModel,
@@ -273,10 +272,11 @@ describe('The Sync Granules workflow', () => {
     let oldExecution;
     let oldUpdatedAt;
     let reingestResponse;
+    let syncGranuleTaskOutput;
     let granule;
 
     beforeAll(async () => {
-      const granuleResponse = await granulesApiTestUtils.getGranule({
+      const granuleResponse = await getGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId,
       });
@@ -284,11 +284,25 @@ describe('The Sync Granules workflow', () => {
 
       oldUpdatedAt = granule.updatedAt;
       oldExecution = granule.execution;
-      const reingestGranuleResponse = await granulesApiTestUtils.reingestGranule({
+      const reingestGranuleResponse = await reingestGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId,
       });
       reingestResponse = JSON.parse(reingestGranuleResponse.body);
+    });
+
+    afterAll(async () => {
+      await deleteGranule({
+        prefix: config.stackName,
+        granuleId: reingestResponse.granuleId,
+      });
+
+      await Promise.all(syncGranuleTaskOutput.payload.granules.map(
+        (g) => deleteGranule({
+          prefix: config.stackName,
+          granuleId: g.granuleId,
+        })
+      ));
     });
 
     it('executes successfully', () => {
@@ -316,7 +330,7 @@ describe('The Sync Granules workflow', () => {
 
       await waitForCompletedExecution(reingestGranuleExecutionArn);
 
-      const syncGranuleTaskOutput = await lambdaStep.getStepOutput(
+      syncGranuleTaskOutput = await lambdaStep.getStepOutput(
         reingestGranuleExecutionArn,
         'SyncGranule'
       );
@@ -331,7 +345,7 @@ describe('The Sync Granules workflow', () => {
         'completed'
       );
 
-      const updatedGranuleResponse = await granulesApiTestUtils.getGranule({
+      const updatedGranuleResponse = await getGranule({
         prefix: config.stackName,
         granuleId: inputPayload.granules[0].granuleId,
       });
@@ -362,9 +376,17 @@ describe('The Sync Granules workflow', () => {
       failingExecution = await buildAndExecuteWorkflow(
         config.stackName, config.bucket, workflowName, collection, provider, inputPayload
       );
+      failingExecutionArn = failingExecution.executionArn;
+
       lambdaOutput = await lambdaStep.getStepOutput(failingExecution.executionArn, 'SyncGranule', 'failure');
 
-      failingExecutionArn = failingExecution.executionArn;
+      // Immediately clean up granules
+      await Promise.all(inputPayload.granules.map(
+        (granule) => deleteGranule({
+          prefix: config.stackName,
+          granuleId: granule.granuleId,
+        })
+      ));
     });
 
     it('completes execution with failure status', () => {